fix(geometry): report worst-host bbox in world space, not the local frame
The per-host `bbox` diagnostic captured `mesh.bounds()` inside
apply_void_context_inner, but on the wasm/local-frame default the host is
stored relative to a nonzero `mesh.origin` and apply_void_context clears that
origin before the inner call — so the captured bounds were the near-zero local
box, not world space, misdirecting `diagnose-geometry --product/--type` and
georeferenced models (the local-frame-consumer bug class: every world-space
reader must fold MeshData.origin).

Capture the world-space host AABB once in apply_void_context, before the
origin clear and before try_cut_wall_local_frame rotates the mesh into its own
frame, and thread it down to both record_host_cut_effect call sites (fast path
and exact path). Origin is folded in f64 so georeferenced hosts keep as much
precision as the f32 diagnostic surface allows.

Adds a pure unit test (world_host_bounds_folds_the_origin_back_in) and an
end-to-end regression test (nonzero_origin_host_records_world_space_bbox) that
drives a nonzero-origin host through apply_void_context and asserts the drained
diagnostic reports world coords. Extracts the flap_clip_tests module into its
own file (mirroring the sibling reveal_tests) so voids/mod.rs stays under its
module-size ratchet budget (shrinks 1963 -> 1889).
