+++ rust/geometry/src/router/voids/mod.rs
@@ -442,6 +442,27 @@ fn recut_malformed_openings(result: &mut Mesh, boxes: &[OpeningBox]) {
 /// collapse on the coarse world grid (#1310 review). A world-framed cutter
 /// (`origin == 0` — the native and ≤100-vertex default) reduces to the plain
 /// `position - host_origin` and stays byte-identical.
+/// World-space AABB of a host mesh: its local `bounds()` with `mesh.origin`
+/// folded back in (world = origin + position). Folded in f64 so a
+/// georeferenced host (large origin) keeps as much precision as the f32
+/// diagnostic surface allows, rather than reporting near-zero local coords.
+fn world_host_bounds(mesh: &Mesh) -> ((f32, f32, f32), (f32, f32, f32)) {
+    let o = mesh.origin;
+    let (mn, mx) = mesh.bounds();
+    (
+        (
+            (mn.x as f64 + o[0]) as f32,
+            (mn.y as f64 + o[1]) as f32,
+            (mn.z as f64 + o[2]) as f32,
+        ),
+        (
+            (mx.x as f64 + o[0]) as f32,
+            (mx.y as f64 + o[1]) as f32,
+            (mx.z as f64 + o[2]) as f32,
+        ),
+    )
+}
+
 fn translate_cutter_mesh(mesh: &Mesh, host_origin: [f64; 3]) -> Mesh {
     let o = mesh.origin;
     let mut positions = Vec::with_capacity(mesh.positions.len());
@@ -853,10 +874,22 @@ impl GeometryRouter {
             }
         }
 
+        // WORLD-space host AABB for the per-host diagnostic (`bbox`), captured
+        // HERE — before the local-frame branch below zeroes `mesh.origin` and
+        // before `try_cut_wall_local_frame` rotates the mesh into its own frame.
+        // On the wasm/local-frame default the host is stored relative to a
+        // nonzero `mesh.origin` (world = origin + position), so bounds taken
+        // after the clear would be local/near-zero and misdirect
+        // `diagnose-geometry --product/--type` (the #1474 local-frame-consumer
+        // bug class: every world-space reader must fold `MeshData.origin`). The
+        // origin is folded in f64 so georeferenced hosts report true world
+        // coords, then cast to the f32 diagnostic surface.
+        let host_world_bounds = world_host_bounds(&mesh);
+
         let origin = mesh.origin;
         if origin == [0.0, 0.0, 0.0] && ctx.all_cutters_world_framed() {
             // Legacy/world frame on host AND cutters: no relativization needed.
-            return self.apply_void_context_inner(mesh, ctx, element_id);
+            return self.apply_void_context_inner(mesh, ctx, element_id, host_world_bounds);
         }
         // Work entirely in the host's local frame (origin 0 on every operand).
         // `relativized_by` folds each cutter's OWN origin and subtracts the host
@@ -865,7 +898,8 @@ impl GeometryRouter {
         // openings did not (origin == 0 but cutters are local-framed).
         mesh.origin = [0.0, 0.0, 0.0];
         let local_ctx = ctx.relativized_by(origin);
-        let mut result = self.apply_void_context_inner(mesh, &local_ctx, element_id);
+        let mut result =
+            self.apply_void_context_inner(mesh, &local_ctx, element_id, host_world_bounds);
         result.origin = origin;
         result
     }
@@ -926,6 +960,7 @@ impl GeometryRouter {
         mesh: &Mesh,
         ctx: &VoidContext,
         element_id: u32,
+        host_world_bounds: ((f32, f32, f32), (f32, f32, f32)),
     ) -> Option<Mesh> {
         if ctx.merged_openings.is_empty() {
             return None;
@@ -1022,23 +1057,30 @@ impl GeometryRouter {
             param: None,
         };
 
-        let result_local = self.apply_void_context_inner(host_local, &local_ctx, element_id);
+        // Forward the WORLD host bounds captured before this rotation so the
+        // diagnostic reports world coords, not wall-frame (rotated/centred) ones.
+        let result_local =
+            self.apply_void_context_inner(host_local, &local_ctx, element_id, host_world_bounds);
         Some(mesh_from_frame(&result_local, &axes, center))
     }
 
     // `host_mutated` is set just before an early `break`, so the final write is
     // intentionally never read back; keep the flag for readability of the branch.
     #[allow(unused_assignments)]
-    fn apply_void_context_inner(&self, mesh: Mesh, ctx: &VoidContext, element_id: u32) -> Mesh {
-        // Capture the input triangle count + bounds so the per-host
-        // diagnostic can flag the "cuts attempted but produced no
-        // change" case — the silent-no-op signature when an opening
-        // box doesn't intersect the host mesh.
+    fn apply_void_context_inner(
+        &self,
+        mesh: Mesh,
+        ctx: &VoidContext,
+        element_id: u32,
+        host_bounds_capture: ((f32, f32, f32), (f32, f32, f32)),
+    ) -> Mesh {
+        // Capture the input triangle count so the per-host diagnostic can flag
+        // the "cuts attempted but produced no change" case — the silent-no-op
+        // signature when an opening box doesn't intersect the host mesh. The
+        // WORLD-space host AABB (`host_bounds_capture`) is passed in from
+        // `apply_void_context` (captured before the origin clear / frame
+        // rotation), not recomputed here from the possibly local-framed `mesh`.
         let tris_before = mesh.triangle_count();
-        let host_bounds_capture = {
-            let (mn, mx) = mesh.bounds();
-            ((mn.x, mn.y, mn.z), (mx.x, mx.y, mx.z))
-        };
         if ctx.is_noop() {
             return mesh;
         }
@@ -1049,7 +1091,8 @@ impl GeometryRouter {
         // The world-space tilted cut at large coordinates over-cuts and
         // fragments badly. Scoped to plan-rotated walls; everything else falls
         // through to the world path unchanged.
-        if let Some(cut) = self.try_cut_wall_local_frame(&mesh, ctx, element_id) {
+        if let Some(cut) = self.try_cut_wall_local_frame(&mesh, ctx, element_id, host_bounds_capture)
+        {
             return cut;
         }
 
@@ -1843,121 +1886,4 @@ impl GeometryRouter {
 }
 
 #[cfg(test)]
-mod flap_clip_tests {
-    use super::*;
-
-    fn box_cutter_mesh(half: [f64; 3], garbage: &[[f64; 3]]) -> Mesh {
-        // 8 corners of an axis-aligned box centred at origin + far garbage verts.
-        let mut m = Mesh::new();
-        for sx in [-1.0, 1.0] {
-            for sy in [-1.0, 1.0] {
-                for sz in [-1.0, 1.0] {
-                    m.positions.extend_from_slice(&[
-                        (sx * half[0]) as f32,
-                        (sy * half[1]) as f32,
-                        (sz * half[2]) as f32,
-                    ]);
-                }
-            }
-        }
-        for g in garbage {
-            m.positions
-                .extend_from_slice(&[g[0] as f32, g[1] as f32, g[2] as f32]);
-        }
-        m
-    }
-
-    /// A cutter with far garbage "fins" is detected as malformed and its real
-    /// opening box is recovered (the fins are excluded).
-    #[test]
-    fn opening_obb_detects_malformed_and_recovers_box() {
-        let cutter = box_cutter_mesh([1.0, 0.1, 1.2], &[[0.0, 9.0, 0.0], [0.0, -9.0, 0.0]]);
-        let b = opening_obb_if_malformed(&cutter).expect("malformed cutter -> box");
-        let mut half = b.half;
-        half.sort_by(|a, c| a.partial_cmp(c).unwrap());
-        assert!((half[0] - 0.1).abs() < 0.05, "thin half {:?}", b.half);
-        assert!((half[1] - 1.0).abs() < 0.05, "mid half {:?}", b.half);
-        assert!((half[2] - 1.2).abs() < 0.05, "long half {:?}", b.half);
-    }
-
-    /// A well-formed cutter (no far cluster) is NOT reshaped.
-    #[test]
-    fn opening_obb_skips_wellformed_cutter() {
-        let cutter = box_cutter_mesh([1.0, 0.1, 1.2], &[]);
-        assert!(opening_obb_if_malformed(&cutter).is_none());
-    }
-
-    /// A watertight TALL cutter — a roof/gable void authored as a closed prism
-    /// reaching far up (here ~900 m) to clip a wall down to the roofline — is a
-    /// VALID SOLID, so its far (top) vertices are STRUCTURAL, not garbage. The
-    /// closed-manifold gate must spare it: flagging it malformed re-cut every
-    /// roof-capped wall as a flat horizontal slab, slicing the gable off (the
-    /// #1440 false-positive). `aabb_box` welds to a closed box, mirroring the
-    /// `IfcClosedShell` roof cutters that triggered the regression.
-    #[test]
-    fn watertight_tall_roof_cutter_is_not_flagged_malformed() {
-        let tall = aabb_box([0.6, 0.15, 450.0]);
-        assert!(
-            cutter_is_closed_manifold(&tall),
-            "a closed box prism must read as a valid solid"
-        );
-        assert!(
-            opening_obb_if_malformed(&tall).is_none(),
-            "a watertight tall cutter must never be reshaped as 'malformed'"
-        );
-    }
-
-    fn signed_volume(m: &Mesh) -> f64 {
-        let v = |i: u32| {
-            let b = i as usize * 3;
-            [m.positions[b] as f64, m.positions[b + 1] as f64, m.positions[b + 2] as f64]
-        };
-        m.indices
-            .chunks_exact(3)
-            .map(|t| {
-                let (a, b, c) = (v(t[0]), v(t[1]), v(t[2]));
-                (a[0] * (b[1] * c[2] - b[2] * c[1]) + a[1] * (b[2] * c[0] - b[0] * c[2])
-                    + a[2] * (b[0] * c[1] - b[1] * c[0]))
-                    / 6.0
-            })
-            .sum::<f64>()
-            .abs()
-    }
-
-    /// An axis-aligned box mesh (helper) via the canonical corner order.
-    fn aabb_box(half: [f64; 3]) -> Mesh {
-        let axes = [Vector3::x(), Vector3::y(), Vector3::z()];
-        let bx = OpeningBox { center: Vector3::zeros(), axes, half };
-        bx.extended_box_mesh([0.0; 3], 0.0)
-    }
-
-    /// `recut_malformed_openings` carves a clean through-opening AND preserves
-    /// the wall AROUND it — the regression where a plain triangle-drop also
-    /// removed the legitimate wall above/below the opening.
-    #[test]
-    fn recut_carves_opening_and_preserves_wall_around_it() {
-        // Solid wall box: 4 (x) x 0.3 (y) x 3 (z) centred at origin.
-        let mut host = aabb_box([2.0, 0.15, 1.5]);
-        let host_vol = signed_volume(&host);
-        // A 1 x 1 window through it (thin axis y; recut extends it through).
-        let bx = OpeningBox {
-            center: Vector3::zeros(),
-            axes: [Vector3::x(), Vector3::y(), Vector3::z()],
-            half: [0.5, 0.079, 0.5],
-        };
-        recut_malformed_openings(&mut host, std::slice::from_ref(&bx));
-        assert!(!host.is_empty(), "recut emptied the wall");
-        // Wall extent preserved on every face axis (no over-cut of the wall
-        // above/below/beside the opening).
-        let (lo, hi) = host.bounds();
-        assert!((hi.z - 1.5).abs() < 0.02, "wall top removed (z max {})", hi.z);
-        assert!((lo.z + 1.5).abs() < 0.02, "wall bottom removed (z min {})", lo.z);
-        assert!((hi.x - 2.0).abs() < 0.02, "wall side removed (x max {})", hi.x);
-        // The opening prism (~1 x 0.3 x 1 = 0.3 m^3) was actually carved out.
-        let cut_vol = signed_volume(&host);
-        assert!(
-            cut_vol < host_vol - 0.2,
-            "opening not carved (host {host_vol:.3}, cut {cut_vol:.3})"
-        );
-    }
-}
+mod flap_clip_tests;
